requirements.txt
# FastAPI and related
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=12.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.12.0

# Pydantic for data validation (using compatible versions)
pydantic>=2.4.0,<2.6.0
pydantic-settings>=2.0.0

# Environment management
python-dotenv>=1.0.0

# CORS and middleware
fastapi-cors>=0.0.6

# Date/time handling
python-dateutil>=2.8.0

# HTTP client for health checks
httpx>=0.25.0

# LangChain and LangGraph (as per requirements)
langchain-core>=0.1.0
langchain-google-genai>=0.1.0
langgraph>=0.1.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Development tools
black>=23.0.0
flake8>=6.0.0